Vector

Description

functions wich connected to vector calculations

Documentation

getColor :: (Point3D, Point3D, Point3D) -> Point3D -> Color

choice of color according to relation to light vetor

drawOne :: Array Int Point3D -> Point3D -> Int -> Triangle -> Picture

Draw on triangle. If triangle is in the back of figure it will not be painted. Color of figure depends on position of triangle in relation to light vetor

rotatePointXZ :: Float -> Point3D -> Point3D

rotate point around Y0

rotatePointYZ :: Float -> Point3D -> Point3D

rotate point around X0

scalePoint :: Float -> Point3D -> Point3D

increase lenth of vector to point

convertPoint :: Int -> Point3D -> (Float, Float)

convert point to coordinate of display